+++ src/modules/header.js
@@ -5,15 +5,36 @@ const Header = () => {
     itemCounter(document);
   });
 
+  window.addEventListener('load', () => {
+    const moviesLink = document.querySelector('.movie-counter');
+    const lastestLink = document.querySelector('.last-movies');
+
+    moviesLink.addEventListener('click', () => {
+      window.dispatchEvent(new CustomEvent('load_content', {
+        detail: {
+          home: true,
+        },
+      }));
+    });
+
+    lastestLink.addEventListener('click', () => {
+      window.dispatchEvent(new CustomEvent('load_content', {
+        detail: {
+          home: false,
+        },
+      }));
+    });
+  });
+
   return `<div id="logo"></div>
   <span class = "nav-menu open"><svg class="svg-inline--fa fa-bars fa-3x" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="bars" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512" data-fa-i2svg=""><path fill="currentColor" d="M0 96C0 78.3 14.3 64 32 64H416c17.7 0 32 14.3 32 32s-14.3 32-32 32H32C14.3 128 0 113.7 0 96zM0 256c0-17.7 14.3-32 32-32H416c17.7 0 32 14.3 32 32s-14.3 32-32 32H32c-17.7 0-32-14.3-32-32zM448 416c0 17.7-14.3 32-32 32H32c-17.7 0-32-14.3-32-32s14.3-32 32-32H416c17.7 0 32 14.3 32 32z"></path></svg></span>
   <span class = "nav-menu close"><svg class="svg-inline--fa fa-xmark fa-3x" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="xmark" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" data-fa-i2svg=""><path fill="currentColor" d="M310.6 150.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L160 210.7 54.6 105.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L114.7 256 9.4 361.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L160 301.3 265.4 406.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L205.3 256 310.6 150.6z"></path></svg></span>
   <nav class="navbar">
      <ul class = "list">
-       <li><a href="" class="movie-counter">Movies</a></li>
-       <li><a href="">Latest</a></li>
-       <li><a href="">Favorite</a></li>
-       <li><a href="">Upcoming</a></li>
+       <li><a  class="movie-counter">Movies</a></li>
+       <li><a class="last-movies">Latest</a></li>
+       <li><a >Favorite</a></li>
+       <li><a >Upcoming</a></li>
      </ul>
    </nav>`;
 };
