+++ src/styles/global.css
@@ -63,6 +63,7 @@
 
 .navbar li {
   list-style: none;
+  cursor: pointer;
 }
 
 .navbar a {
